requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chirptext==0.1.2
click==8.1.8
cryptography==46.0.7
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
future @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jamdict==0.1a11.post2
jiter==0.14.0
macholib @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
openai==2.31.0
proto-plus==1.27.2
protobuf==5.29.6
puchikarui==0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.1
pydantic_core==2.46.1
pyparsing==3.3.2
requests==2.32.5
six @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
